Refactor CLIP models.
